Add a per-thread struct task cache.
This allows us to save some malloc()/free() calls.

The structs can leak on program exit, but an external vector to store
them could fix this.

As it stands, this isn't really useful: in multiple tests with the linux
kernel tree, the p_old branch was taken between 0 and 3 times in a full
run.