Add directory file descriptor caching as well.
Now a parent can store its own fd, which will live on for as long as the
parent does, allowing us to use relative *at functions (that might be
faster on the kernel side?) and store smaller path buffers.

Since we were playing with soft limit detection, also make the EMFILE
operations conditional on limited_fds, avoiding potentially expensive
mutex operations when possible.

When checking the soft limit, we use 2 for the number of standard
file descriptors, because we closed stdin in main().